Use BigDecimal for money & add Android logger
Migrate monetary values from double/Double to BigDecimal across the Android UI and ViewModel layers to improve precision. Adjusted adapters, fragments and viewmodels to use BigDecimal (and to call doubleValue() where needed for formatting/display). Added AndroidAppLogger and wired it into AppContext/CoreBootstrap to provide platform logging. Also removed the deprecated getRegistrarLancamentoParceladoCartaoUseCase accessor and updated related calculations and chart/report code to operate with BigDecimal totals.

+++ android/src/main/java/com/financeiramente/android/ui/cartao/CartaoAdapter.java
@@ -82,7 +82,7 @@ public class CartaoAdapter extends RecyclerView.Adapter<CartaoAdapter.ViewHolder
             tvVencimento.setText(String.format(Locale.getDefault(),
                     "Vencimento: dia %d", cartao.getDiaVencimento()));
 
-            if (cartao.getLimite() != null && cartao.getLimite() > 0) {
+            if (cartao.getLimite() != null && cartao.getLimite().doubleValue() > 0) {
                 tvLimiteTotal.setText(String.format(Locale.getDefault(), "Limite: R$ %.2f", cartao.getLimite()));
                 tvLimiteTotal.setVisibility(View.VISIBLE);
                 tvUtilizado.setVisibility(View.VISIBLE);
