+++ core/src/main/java/com/financeiramente/core/usecase/cartao/CriarCartaoCreditoInput.java
@@ -1,19 +1,20 @@
 package com.financeiramente.core.usecase.cartao;
 
 import com.financeiramente.core.domain.vo.BandeiraCartao;
+import java.math.BigDecimal;
 import java.util.List;
 
 public class CriarCartaoCreditoInput {
     private final String nome;
     private final int diaVencimento;
     private final int diasParaFechamento;
-    private final Double limite;
+    private final BigDecimal limite;
     private final BandeiraCartao bandeira;
     private final String icone;
     private final String cor;
 
     public CriarCartaoCreditoInput(String nome, int diaVencimento, int diasParaFechamento,
-                                   Double limite, BandeiraCartao bandeira,
+                                   BigDecimal limite, BandeiraCartao bandeira,
                                    String icone, String cor) {
         this.nome = nome;
         this.diaVencimento = diaVencimento;
@@ -27,7 +28,7 @@ public class CriarCartaoCreditoInput {
     public String getNome() { return nome; }
     public int getDiaVencimento() { return diaVencimento; }
     public int getDiasParaFechamento() { return diasParaFechamento; }
-    public Double getLimite() { return limite; }
+    public BigDecimal getLimite() { return limite; }
     public BandeiraCartao getBandeira() { return bandeira; }
     public String getIcone() { return icone; }
     public String getCor() { return cor; }
