+++ core/src/main/java/com/financeiramente/core/usecase/fatura/PagamentoFaturaResult.java
@@ -1,17 +1,18 @@
 package com.financeiramente.core.usecase.fatura;
 
 import com.financeiramente.core.domain.vo.StatusFatura;
+import java.math.BigDecimal;
 
 public class PagamentoFaturaResult {
 
     private final StatusFatura statusResultante;
-    private final double saldoDevedor;
+    private final BigDecimal saldoDevedor;
 
-    public PagamentoFaturaResult(StatusFatura statusResultante, double saldoDevedor) {
+    public PagamentoFaturaResult(StatusFatura statusResultante, BigDecimal saldoDevedor) {
         this.statusResultante = statusResultante;
         this.saldoDevedor = saldoDevedor;
     }
 
     public StatusFatura getStatusResultante() { return statusResultante; }
-    public double getSaldoDevedor() { return saldoDevedor; }
+    public BigDecimal getSaldoDevedor() { return saldoDevedor; }
 }
